Module x509_cert::certificate

Certificate types

Structs

CertificateInner

X.509 certificates are defined in RFC 5280 Section 4.1.

Rfc5280

Parse certificates with rfc5280-compliant checks

TbsCertificateInner

X.509 TbsCertificate as defined in RFC 5280 Section 4.1

Enums

Version

Certificate Version as defined in RFC 5280 Section 4.1.

Traits

Profile

Profile allows the consumer of this crate to customize the behavior when parsing certificates. By default, parsing will be made in a rfc5280-compliant manner.

Type Definitions

Certificate

X.509 certificates are defined in RFC 5280 Section 4.1.

PkiPath

PkiPath as defined by X.509 and referenced by RFC 6066.

TbsCertificate

X.509 TbsCertificate as defined in [RFC 5280 Section 4.1]